Remove visitor trait
