+++ src/Language/Node/FieldNode.php
@@ -4,8 +4,12 @@ namespace Digia\GraphQL\Language\Node;
 
 use Digia\GraphQL\Language\Location;
 
-class FieldNode extends AbstractNode implements SelectionNodeInterface, ArgumentsAwareInterface,
-    DirectivesAwareInterface, NameAwareInterface, SelectionSetAwareInterface
+class FieldNode extends AbstractNode implements
+    SelectionNodeInterface,
+    NameAwareInterface,
+    ArgumentsAwareInterface,
+    DirectivesAwareInterface,
+    SelectionSetAwareInterface
 {
     use NameTrait;
     use AliasTrait;
