+++ src/Language/Node/NodeInterface.php
@@ -3,6 +3,7 @@
 namespace Digia\GraphQL\Language\Node;
 
 use Digia\GraphQL\Language\Location;
+use Digia\GraphQL\Language\Visitor\VisitorInterface;
 
 interface NodeInterface
 {
@@ -11,6 +12,11 @@ interface NodeInterface
      * @return string
      */
     public function getKind(): string;
+
+    /**
+     * @return bool
+     */
+    public function hasLocation(): bool;
 
     /**
      * @return Location|null
@@ -26,4 +32,36 @@ interface NodeInterface
      * @return string
      */
     public function toJSON(): string;
+
+    /**
+     * @param VisitorInterface   $visitor
+     * @param string|int         $key
+     * @param NodeInterface|null $parent
+     * @param array              $path
+     * @param array              $ancestors
+     * @return NodeInterface|null
+     */
+    public function acceptVisitor(
+        VisitorInterface $visitor,
+        $key = null,
+        ?NodeInterface $parent = null,
+        array $path = [],
+        array $ancestors = []
+    ): ?NodeInterface;
+
+    /**
+     * @return NodeInterface|null
+     */
+    public function getAncestor(int $depth = 1): ?NodeInterface;
+
+    /**
+     * @param NodeInterface $node
+     * @return bool
+     */
+    public function determineIsEdited(NodeInterface $node): bool;
+
+    /**
+     * @return bool
+     */
+    public function isEdited(): bool;
 }
