+++ src/Language/Node/UnionTypeExtensionNode.php
@@ -4,8 +4,10 @@ namespace Digia\GraphQL\Language\Node;
 
 use Digia\GraphQL\Language\Location;
 
-class UnionTypeExtensionNode extends AbstractNode implements TypeSystemExtensionNodeInterface, DirectivesAwareInterface,
-    NameAwareInterface
+class UnionTypeExtensionNode extends AbstractNode implements
+    TypeSystemExtensionNodeInterface,
+    NameAwareInterface,
+    DirectivesAwareInterface
 {
     use NameTrait;
     use DirectivesTrait;
